LayoutTests/imported/w3c/web-platform-tests/css/css-transitions/starting-style-size-container-expected.txt


FAIL Triggered transition from first style update based on up-to-date container query assert_equals: @starting-style based on the size query evaluation from the same frame expected "rgb(128, 255, 128)" but got "rgb(0, 255, 0)"
